requirements.txt
# Database clients
psycopg2-binary>=2.9.9
pgvector>=0.2.4
qdrant-client>=1.7.0
weaviate-client>=4.4.0
pymilvus>=2.3.0
chromadb>=0.4.15
faiss-cpu>=1.7.4
opensearch-py>=2.3.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Embeddings and ML
openai>=1.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Document parsing
PyPDF2>=3.0.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Progress and logging
tqdm>=4.65.0
rich>=13.0.0

# System monitoring
psutil>=5.9.0

# HTTP and API
requests>=2.31.0
httpx>=0.24.0
